fix: Package update and minor fix
Updated the reference to `Microsoft.EntityFrameworkCore.Design` to version `8.0.6` in the `TccRewritenCsharp.Infrastructure.csproj` file, including specific settings for `PrivateAssets` and `IncludeAssets`. The version of the `Microsoft.EntityFrameworkCore.Sqlite` package has also been updated from `8.0.5` to `8.0.6`. In the file `TccRewritenCsharp.Test.csproj`, a change was made that suggests a coding correction or the addition of a BOM, without visible changes to the code content.